+++ app/src/main/java/com/example/cardinalgate/MainActivity.java
@@ -2,7 +2,6 @@ package com.example.cardinalgate;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 
@@ -15,9 +14,7 @@ import androidx.navigation.ui.NavigationUI;
 
 import com.example.cardinalgate.core.TokenManager;
 import com.example.cardinalgate.core.UserDataManager;
-import com.example.cardinalgate.core.api.model.responses.SummaryResponse;
 import com.example.cardinalgate.databinding.ActivityMainBinding;
-import com.example.cardinalgate.ui.home.HomeFragment;
 import com.google.android.material.navigation.NavigationView;
 
 public class MainActivity extends AppCompatActivity {