+++ app/src/main/java/com/example/cardinalgate/core/api/model/calls/AddCardCall.java
@@ -1,0 +1,8 @@
+package com.example.cardinalgate.core.api.model.calls;
+
+import com.google.gson.annotations.SerializedName;
+
+public class AddCardCall {
+    @SerializedName("cardId")
+    public String cardId;
+}