+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Update monochrome icon
 - Add context menu to home screen mascot
 - Improve carousel text in light mode
+- Implement cards management
 
 ## [1.6]
 - Add keystore to project to sign the APK